+++ lib/quo/composing/query_classes_strategy.rb
@@ -28,7 +28,7 @@ module Quo
       # @rbs return: Class & Quo::ComposedQuery
       def compose(chosen_superclass, left_query_class, right_query_class, joins: nil, left_spec: nil, right_spec: nil)
         validate_query_classes(left_query_class, right_query_class)
-        props = collect_properties(left_query_class, right_query_class)
+        props = collect_properties(chosen_superclass, left_query_class, right_query_class)
         klass = create_composed_class(chosen_superclass, props)
         assign_query_metadata(klass, left_query_class, right_query_class, joins, left_spec, right_spec)
         klass